apps/forge-lite/api/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
httpx>=0.25.0
requests>=2.31.0
